requirements.txt
ib_insync
pandas~=2.3.3
ta~=0.11.0
pydantic-settings~=2.11.0
boto3~=1.40.71
ibapi~=10.37.2
requests~=2.32.5
fastapi~=0.120.0
setuptools~=80.9.0
pydantic~=2.12.3
numpy~=2.3.4
PyYAML~=6.0.3